arcappcompat: Update ClickUntilButtonExists func in loginutil file
Bloxel is failing due to “Failed to restore the window:” error since the allow button is not clicked when the dialog box to allow app to take pictures and  photos” appears. Added a wait time for the allow button to exist. Updated the ClickUntilButtonExists func in the “loginutil.go” file. This will help to fix the failures.

BUG=b:253082101
TEST=tast run <zork-arc-r> arcappcompat.Bloxels.vm_clamshell_mode_top_apps

Change-Id: I0924454ba3670c0d2da5474ed48507b739a4dccb